Rename adapter util file and move `AdapterDayjs` undefined locale warning
