+++ codecov.yml
@@ -2,7 +2,7 @@ coverage:
   ignore:
     - '**/*.test.tsx'
     # ignore date adapter parts that are hardly testable
-    - '**/assertDateAdapters.ts'
+    - '**/internals/utils/adapters.ts'
   status:
     project:
       default:
