+++ packages/x-date-pickers/src/AdapterDayjs/AdapterDayjs.ts
@@ -13,7 +13,7 @@ import {
   PickersTimezone,
   DateBuilderReturnType,
 } from '../models';
-import { warnOnce } from '../internals/utils/warning';
+import { warnUndefinedDayjsLocale } from '../internals/utils/adapters';
 
 defaultDayjs.extend(localizedFormatPlugin);
 defaultDayjs.extend(weekOfYearPlugin);
@@ -248,14 +248,7 @@ export class AdapterDayjs implements MuiPickersAdapter<Dayjs, string> {
     let localeObject = locales[locale];
 
     if (localeObject === undefined) {
-      if (process.env.NODE_ENV !== 'production') {
-        warnOnce([
-          'MUI X: Your locale has not been found.',
-          'Either the locale key is not a supported one. Locales supported by dayjs are available here: https://github.com/iamkun/dayjs/tree/dev/src/locale.',
-          "Or you forget to import the locale from 'dayjs/locale/{localeUsed}'",
-          'fallback on English locale.',
-        ]);
-      }
+      warnUndefinedDayjsLocale();
       localeObject = locales.en;
     }
